Extracted base functionality out of ResourceHandler into AbstractResourceHandler base for more flexible use.
